rename SSUtxoFlow to SSSankeyDiagram
